requirements/base.txt
albumentations
filelock
hydra-core>=0.11,<1.0
hydra-colorlog
numpy
omegaconf>=1.4,<2.0
Pillow
scipy
scikits.bootstrap>=1.0,<1.1
tensorflow>=2.2.0,<2.3.0
